Improve Ban Commands
Fix issue #121 where the ban reason was empty. This was due to improper string processing.

Stop processing multiple arguments which is really messy. Require the user to put the ban reason in quotes.

Include "N/A" as the default ban reason, so it's clearly stated when users do not input a reason.