Remove StripQuotes
StripQuotes is very useless. The game plugin already strips it, so it's like \". The problem is when somebody has quotes on end of his name - then it removes the quote and there is only back slash for the function. So it can looks like (when somebody has for example Test Name"):
RemoveBlock('1234', '...', '', 'Test Name\', 0);