Fix cookie domain for non-standard ports.
Previously the cookie domain would be set to whatever the httpd's
HTTP_HOST is. But in the case of web servers using non-standard ports,
eg: foo.bar.com:8080, this would cause the cookie domain to also be set
to foo.bar.com:8080.

The problem is that cookie domains do not account for ports, and
subsequent requests will not send the cookie back because the domain
that the browser wants is just `foo.bar.com`, not `foo.bar.com:8080`.

This commit adds a simple `Host::cookieDomain()` wrapper around
`Host::domain()`, and replaces the relevant usages in the Auth class.